Refactor dashboard header components
JIRA: RAIL-4183
