+++ libs/sdk-ui-dashboard/.dependency-cruiser.js
@@ -35,6 +35,7 @@ options = {
                 "src/presentation/localization",
                 "src/presentation/scheduledEmail",
                 "src/presentation/shareDialog",
+                "src/presentation/deleteDialog",
                 "src/presentation/saveAs",
                 "src/presentation/topBar",
                 "src/presentation/widget",
@@ -110,6 +111,7 @@ options = {
             "src/model",
         ]),
         depCruiser.moduleWithDependencies("topBar", "src/presentation/topBar", [
+            "src/_staging/*",
             "src/presentation/dashboardContexts",
             "src/presentation/localization",
             "src/presentation/constants/*",
