+++ libs/sdk-ui-dashboard/src/presentation/filterBar/filterBar/index.tsx
@@ -1,5 +1,5 @@
-// (C) 2021 GoodData Corporation
-export { DefaultFilterBar } from "./DefaultFilterBar";
+// (C) 2021-2022 GoodData Corporation
+export { DefaultFilterBar, useFilterBarProps } from "./DefaultFilterBar";
 export { FilterBar } from "./FilterBar";
 export { HiddenFilterBar } from "./HiddenFilterBar";
 export * from "./types";
