+++ libs/sdk-ui-dashboard/src/presentation/saveAs/index.ts
@@ -1,4 +1,4 @@
-// (C) 2020-2021 GoodData Corporation
-export { DefaultSaveAsDialog } from "./DefaultSaveAsDialog";
+// (C) 2020-2022 GoodData Corporation
+export { DefaultSaveAsDialog, useSaveAsDialogProps } from "./DefaultSaveAsDialog";
 export { SaveAsDialog } from "./SaveAsDialog";
 export { CustomSaveAsDialogComponent, ISaveAsDialogProps } from "./types";
