+++ libs/sdk-ui-dashboard/src/presentation/topBar/menuButton/index.tsx
@@ -1,5 +1,6 @@
-// (C) 2021 GoodData Corporation
+// (C) 2021-2022 GoodData Corporation
 export { DefaultMenuButton } from "./DefaultMenuButton";
 export { HiddenMenuButton } from "./HiddenMenuButton";
 export { MenuButton } from "./MenuButton";
+export { useDefaultMenuItems } from "./useDefaultMenuItems";
 export * from "./types";
