+++ libs/sdk-ui-dashboard/src/presentation/topBar/topBar/DefaultTopBar.tsx
@@ -1,14 +1,79 @@
-// (C) 2021 GoodData Corporation
-import React from "react";
+// (C) 2021-2022 GoodData Corporation
+import React, { useCallback } from "react";
 
-import { selectIsExport, selectLocale, useDashboardSelector } from "../../../model";
-import { IntlWrapper } from "../../localization";
-import { ButtonBar } from "../buttonBar";
-import { MenuButton } from "../menuButton";
+import {
+    renameDashboard,
+    selectDashboardShareInfo,
+    selectDashboardTitle,
+    selectIsExport,
+    selectIsReadOnly,
+    selectPersistedDashboard,
+    useDashboardDispatch,
+    useDashboardSelector,
+} from "../../../model";
+import { ButtonBar, DefaultButtonBar } from "../buttonBar";
+import { DefaultMenuButton, MenuButton, useDefaultMenuItems } from "../menuButton";
 import { Title } from "../title";
 import { ITopBarProps } from "./types";
 import { HiddenTopBar } from "./HiddenTopBar";
 import { DefaultLockedStatus, DefaultShareStatus } from "../shareIndicators";
+import {
+    useCancelButtonProps,
+    useEditButtonProps,
+    useSaveAsNewButtonProps,
+    useSaveButtonProps,
+    useShareButtonProps,
+} from "../buttonBar/button";
+
+/**
+ * @alpha
+ */
+export const useTopBarProps = (): ITopBarProps => {
+    const dispatch = useDashboardDispatch();
+    const title = useDashboardSelector(selectDashboardTitle);
+    const isReadOnly = useDashboardSelector(selectIsReadOnly);
+    const shareInfo = useDashboardSelector(selectDashboardShareInfo);
+    const persistedDashboard = useDashboardSelector(selectPersistedDashboard);
+
+    const defaultMenuItems = useDefaultMenuItems();
+
+    const shareButtonProps = useShareButtonProps();
+    const editButtonProps = useEditButtonProps();
+    const cancelButtonProps = useCancelButtonProps();
+    const saveButtonProps = useSaveButtonProps();
+    const saveAsNewButtonProps = useSaveAsNewButtonProps();
+
+    const onTitleChanged = useCallback(
+        (title: string) => {
+            dispatch(renameDashboard(title));
+        },
+        [dispatch],
+    );
+
+    return {
+        menuButtonProps: { menuItems: defaultMenuItems, DefaultMenuButton: DefaultMenuButton },
+        titleProps: {
+            title,
+            onTitleChanged: isReadOnly ? undefined : onTitleChanged,
+        },
+        buttonBarProps: {
+            shareButtonProps,
+            editButtonProps,
+            cancelButtonProps,
+            saveButtonProps,
+            saveAsNewButtonProps,
+            DefaultButtonBar: DefaultButtonBar,
+        },
+        shareStatusProps: {
+            shareStatus: shareInfo.shareStatus,
+            isUnderStrictControl: !!persistedDashboard?.isUnderStrictControl,
+        },
+        lockedStatusProps: {
+            isLocked: !!shareInfo.isLocked,
+        },
+        DefaultTopBar,
+    };
+};
 
 const TopBarCore = (props: ITopBarProps): JSX.Element => {
     const { menuButtonProps, titleProps, buttonBarProps, shareStatusProps, lockedStatusProps } = props;
@@ -30,17 +95,12 @@ const TopBarCore = (props: ITopBarProps): JSX.Element => {
 /**
  * @alpha
  */
-export const DefaultTopBar = (props: ITopBarProps): JSX.Element => {
+export function DefaultTopBar(props: ITopBarProps): JSX.Element {
     const isExport = useDashboardSelector(selectIsExport);
-    const locale = useDashboardSelector(selectLocale);
 
     if (isExport) {
         return <HiddenTopBar {...props} />;
     }
 
-    return (
-        <IntlWrapper locale={locale}>
-            <TopBarCore {...props} />
-        </IntlWrapper>
-    );
-};
+    return <TopBarCore {...props} />;
+}
