+++ libs/sdk-ui-dashboard/src/presentation/topBar/topBar/index.tsx
@@ -1,5 +1,5 @@
-// (C) 2021 GoodData Corporation
-export { DefaultTopBar } from "./DefaultTopBar";
+// (C) 2021-2022 GoodData Corporation
+export { DefaultTopBar, useTopBarProps } from "./DefaultTopBar";
 export { HiddenTopBar } from "./HiddenTopBar";
 export { TopBar } from "./TopBar";
 export * from "./types";
